Commit Current State
Trying to work out how to service multiple clients asynchronously.
recv() is blocking by default, so both client and server are stuck
waiting for a new message when no message will be sent. This softlocks
each ClientHandler thread after its first loop. This should be able to
be fixed by setting recv() to be non-blocking or to check if a message
is available before reading.

To-Do:
 - Rewrite sockwrapper.c sockw_read() to check for messages before
   calling recv()
 - Neaten sockwrapper.c sockw_socket() for setting options on fd and
   setsockopt()
 - Ensure both server and client can handle NULL returns from
   sockw_read()